+++ packages/next-common/components/charts/votesBubble/bubbleContent.jsx
@@ -1,0 +1,47 @@
+import clsx from "clsx";
+import User from "next-common/components/user";
+import Tooltip from "next-common/components/tooltip";
+import VoteBubbleHoverContent from "./hoverContent";
+import { useRouter } from "next/router";
+
+// if `d` less than this value, don't render address/identity
+const CONTENT_MIN_SIZE = 60;
+
+export default function VoteBubbleContent({ node, sizeField }) {
+  const router = useRouter();
+  const d = node.r * 2;
+
+  return (
+    <Tooltip
+      className={clsx("!block h-full w-full rounded-full")}
+      content={<VoteBubbleHoverContent node={node} sizeField={sizeField} />}
+    >
+      <div
+        role="link"
+        className={clsx(
+          "flex items-center justify-center cursor-pointer",
+          "rounded-full w-full h-full px-2",
+        )}
+        onClick={() => {
+          router.push(`/user/${node.data.account}/votes`);
+        }}
+      >
+        {d >= CONTENT_MIN_SIZE && (
+          <User
+            add={node.data.account}
+            showAvatar={false}
+            maxWidth={d - 12}
+            noEvent
+            noTooltip
+            ellipsis={false}
+            color={clsx(
+              node.data.aye && "var(--green500)",
+              node.data.aye === false && "var(--red500)",
+              node.data.isAbstain && "var(--textSecondary)",
+            )}
+          />
+        )}
+      </div>
+    </Tooltip>
+  );
+}
